requirements.txt
python-dotenv==1.0.1
matplotlib==3.10.0
pandas==2.2.3
torch==2.5.1
tqdm==4.67.1
scikit-learn==1.6.0